backend/microsservicos/notificacoes/requirements.txt
fastapi
uvicorn
ffmpeg-python
pika
pydantic
pytest
sqlalchemy
passlib[bcrypt]
pymysql
requests
load_dotenv